+++ src/app/beer_garden/db/mongo/models.py
@@ -77,6 +77,8 @@ __all__ = [
     "RemoteRole",
     "User",
     "CommandPublishingBlocklist",
+    "Topic",
+    "Subscriber",
 ]
 
 REQUEST_MAX_PARAM_SIZE = 5 * 1_000_000
@@ -573,6 +575,30 @@ class Request(MongoModel, Document):
                     f"completed status. Requested: {self.status}, completed statuses "
                     f"are {BrewtilsRequest.COMPLETED_STATUSES}."
                 )
+
+
+class Subscriber(MongoModel, EmbeddedDocument):
+    brewtils_model = brewtils.models.Subscriber
+
+    garden = StringField()
+    namespace = StringField()
+    system = StringField()
+    version = StringField()
+    instance = StringField()
+    command = StringField()
+
+
+class Topic(MongoModel, Document):
+    brewtils_model = brewtils.models.Topic
+
+    name = StringField(required=True)
+    subscribers = EmbeddedDocumentListField("Subscriber")
+
+    meta = {
+        "auto_create_index": True,  # We need to manage this ourselves
+        "index_background": True,
+        "indexes": [{"name": "unique_index", "fields": ["name"], "unique": True}],
+    }
 
 
 class System(MongoModel, Document):